move model-independent bivariate normals code to a submodule
